+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 
 - Support for structured custom fields (with fully qualified path)
 
+### Features
+
+- Add sendEmailVerification and verifyEmail
+
 ## v9.5.0 (08/01/2025)
 
 ### Changes